完成顶层模块教程

+++ modules/core/test/test_dataio.cpp
@@ -18,14 +18,14 @@ namespace rm_test
 
 TEST(DataIO, CornersIO)
 {
-    std::vector<std::vector<cv::Point2d>> corners = {{{0.0, 1.1}, {2.2, 3.3}},
+    std::vector<std::vector<cv::Point2f>> corners = {{{0.0, 1.1}, {2.2, 3.3}},
                                                      {{4.4, 5.5}, {6.6, 7.7}, {8.8, 9.9}}};
     EXPECT_TRUE(rm::writeCorners("ts_dataio.yml", 2, corners));
 
     decltype(corners) ret;
     EXPECT_TRUE(rm::readCorners("ts_dataio.yml", 2, ret));
-    EXPECT_EQ(ret[0][0], cv::Point2d(0.0, 1.1));
-    EXPECT_EQ(ret[1][2], cv::Point2d(8.8, 9.9));
+    EXPECT_EQ(ret[0][0], cv::Point2f(0.0, 1.1));
+    EXPECT_EQ(ret[1][2], cv::Point2f(8.8, 9.9));
 }
 
 } // namespace rm_test
